escrow/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-cors-middleware==1.5.0
django-extensions==2.2.6
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
